complete restaurant authorization

+++ src/Restaurants.Application/Commands/Restaurants/DeleteCommands/DeleteRestaurantCommandHandler.cs
@@ -2,7 +2,8 @@
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.CustomExceptions;
 using Restaurants.Domain.Entities;
-using Restaurants.Domain.RepositoryInterfaces;
+using Restaurants.Domain.Enums;
+using Restaurants.Domain.Interfaces;
 
 namespace Restaurants.Application.Commands.Restaurants.DeleteCommands;
 
@@ -10,11 +11,13 @@ public class DeleteRestaurantCommandHandler : IRequestHandler<DeleteRestaurantCo
 {
     private readonly IRestaurantsRepository _restaurantsRepository;
     private readonly ILogger<DeleteRestaurantCommandHandler> _logger;
+    private readonly IRestaurantAuthorizationService _authorizationService;
 
-    public DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, ILogger<DeleteRestaurantCommandHandler> logger)
+    public DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantAuthorizationService authorizationService)
     {
         _restaurantsRepository = restaurantsRepository;
         _logger = logger;
+        _authorizationService = authorizationService;
     }
 
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -25,7 +28,15 @@ public class DeleteRestaurantCommandHandler : IRequestHandler<DeleteRestaurantCo
             var restaurant = await _restaurantsRepository.GetByIdAsync(request.Id)
                 ?? throw new ResourseNotFoundException("Restaurant", request.Id.ToString());
 
+            if(!_authorizationService.Authorize(restaurant, RestaurantOperation.Delete))
+                throw new UnAuthorizedException("User is not authorized to delete this restaurant.");
+
             await _restaurantsRepository.DeleteAsync(restaurant);
+
+        }
+        catch (UnAuthorizedException ex)
+        {
+            throw;
         }
         catch (ResourseNotFoundException ex)
         {
